requirements.txt
streamlit==1.25.0
opencv-python
numpy
streamlit-drawable-canvas
Pillow
